==> 01.ipynb <==
# Cleaning meteo station data

Pro zacatek jsem nasla coordinates oficialnich meterorologickych stanic v ceske republice na

https://www.chmi.cz/aktualni-situace/aktualni-stav-pocasi/ceska-republika/stanice/profesionalni-stanice/tabulky/zemepisne-souradnice?l=en

Zaznamy zahrnuji i jejich nadmorske vysky, coz se bude hodit dal v planovanych analyzach. Data tradicne nevypadaji tak, jak bych si predstavovala, proto jsem si je precistila.

Coordinates budou dale pouzity ke stazeni meteorologickych dat z open meteo.

Imports

import pandas as pd

Functions
- opening and pre-clearing the root csv
- selecting only coordinates
- selecting name, coordinates in float form and elevation data

def open_file(file_path):
    stanice = []
    with open(file_path, mode="r", encoding="utf-8") as my_file:
        for line in my_file:
            line = line.split(",")
            line = [item.replace("°", "") for item in line]
            line = [item.replace("\n", "") for item in line]
            stanice.append(line)
    return stanice

def meteo_souradnice(stanice):
    souradnice = []
    for line in stanice:
        souradnice.append([line[2], line[4], line[6]])
    return souradnice

def seznam_stanic(stanice):
    souradnice = []
    for line in stanice:
        souradnice.append([line[0], line[2], line[4], line[6]])
    return souradnice

Opening and pre-cleaning the data, saving 2 files:
 - csv with meteo-stations coordinates (only) that will be used for downloading data from open meteo
 - csv with cleaned data of the meteo-stations including their names and altitude. This will be used in the future for the analysis.


file_path = "data/01_stanice.csv"
stanice = open_file(file_path)
stanice = stanice[1:]

souradnice = meteo_souradnice(stanice)
df_souradnice = pd.DataFrame(souradnice, columns = ["latitude", "longitude", "elevation"])
df_souradnice.to_csv("data/02_souradnice.csv", index=False, encoding="utf-8")

souradnice = seznam_stanic(stanice)
df_souradnice = pd.DataFrame(souradnice, columns = ["location_id", "latitude", "longitude", "elevation"])
df_souradnice.to_csv("data/03_stanice.csv", encoding="utf-8")
