==> 04_is_true.ipynb <==
Imports

import pandas as pd

katerina = pd.read_csv("data/katerina.csv")
vanoce = pd.read_csv("data/vanoce.csv")
katerina

# this part is under reconstruction. does not run so yet

# katerina.set_index('index', inplace=True)
# vanoce.set_index('index', inplace=True)

# katerina['year'] = pd.to_datetime(katerina['Date']).dt.year
# vanoce['year'] = pd.to_datetime(vanoce['Date']).dt.year

# merged_df = pd.merge(katerina, vanoce, on='index', suffixes=('_katerina', '_vanoce'))

# merged_df['yes'] = merged_df['Temperature_2m_mean_katerina'] > merged_df['Temperature_2m_mean_vanoce']

# result_df = merged_df[['Longtitude_katerina', 'Latitude_katerina', 'Elevation_katerina', 'year', 'yes']]

# result_df.to_csv('data/merged_result.csv', index=False)
